Output image matrix should be uint8

+++ +Subsampling/subsampledToYCbCrImage.m
@@ -18,7 +18,7 @@ end
 
 if isfield(simage, {'y','cb','cr'})
     osize = size(simage.y);
-    ycbcrimage = zeros(osize(1), osize(2), 3);
+    ycbcrimage = uint8(zeros(osize(1), osize(2), 3));
     ycbcrimage(:,:,1) = simage.y;
     ycbcrimage(:,:,2) = imresize(simage.cb, [osize(1) osize(2)], upsampleFilter);
     ycbcrimage(:,:,3) = imresize(simage.cr, [osize(1) osize(2)], upsampleFilter);
